fix(SEO.js): change lang attribute value from 'en' to 'uk' for better localization feat(SEO.js): add support for no
